+++ test/test_helper.rb
@@ -4,5 +4,7 @@ $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 
 require_relative "simplecov_helper"
 require "minitest/autorun"
+require "minitest/mock"
 require "rails"
+require "recording_studio"
 require "recording_studio_notifications_email"